+++ firebase-react-hosting/src/components/TodoForm.js
@@ -2,9 +2,11 @@ import React from 'react'
 import Form from 'muicss/lib/react/form';
 import Button from 'muicss/lib/react/button';
 import Input from 'muicss/lib/react/input';
-import { useState } from 'react';
+import { useState, useContext } from 'react';
+import { TodosContext } from '../contexts/TodoContext';
 
-function TodoForm({addTodo}) {
+function TodoForm() {
+    const { addTodo } = useContext(TodosContext);
     const [value, setValue] = useState('');
  
     const handleSubmit = e => {
@@ -20,6 +22,7 @@ function TodoForm({addTodo}) {
                 label="Input01"
                 floatingLabel={true}
                 defaultValue=""
+                value={value}
                 onChange={(e) => { setValue(e.target.value) }}
             />
             <Button
